refactor: headerUnhide controller
